Add per build type "cargoTest${buildType}" tasks (Closes #67)
In order to support `cargoTest`, we need a way to run the compiled tests on the emulator.
Since we are cross compiling for different architectures, running the binaries directly as `cargo test` does by default will not work.

Thankfully, `cargo test` [reads an environment variable, `CARGO_TARGET_<target_triple>_RUNNER`](https://doc.rust-lang.org/cargo/reference/config.html#targettriplerunner) to control how to run the compiled binaries (including for tests).
This commit introduces a new script that depends on `adb` being available in the path and it already being connected to a running emulator with appropriate architecture.

These new gradle tasks will not work on Windows yet since I am not familiar with `.bat` as needed for that.
A follow-up issue can be filed if there is demand for it.